move config_file fixture to conftest
for easier using in multiple test files
